hw4 和 hw1 小修正
