--- homeworks/week21/hw1/todolist/src/styledcomponent/index.js
@@ -1,3 +1,0 @@
-import { AddTodo, Todos, TodoFilter } from "./todostyled";
-
-export { AddTodo, Todos, TodoFilter };
