+++ homeworks/week21/hw2/hw2/src/App.js
@@ -1,144 +1,6 @@
-import "./index.css";
-import styled from "styled-components";
 import { useState, useEffect } from "react";
-const Container = styled.div`
-  width: 100%;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-`;
-const Block = styled.div`
-  height: 45px;
-  width: 45px;
-  background-color: ${(props) =>
-    props.indexX === 19 || props.indexY === 19 ? "transparent" : "#ccc"};
-  border: 1px solid
-    ${(props) =>
-      props.indexX === 19 || props.indexY === 19 ? "transparent" : "black"};
-  position: relative;
-`;
-const Title = styled.h1``;
-const StateInfo = styled.div`
-  margin-bottom: 20px;
-`;
-const ClickBlock = styled.div`
-  position: absolute;
-  top: -52.5%;
-  left: -52.5%;
-  border: ${(props) => (props.color === null ? "0" : "1px")} solid black;
-  border-radius: 50%;
-  transform: scale(0.75);
-  width: 100%;
-  height: 100%;
-  background-color: ${(props) =>
-    props.color === null ? "transparent" : props.color};
-  :hover {
-    background-color: ${(props) =>
-      props.winner === null
-        ? props.color === null
-          ? props.blackIsNext
-            ? "black"
-            : "white"
-          : props.color
-        : ""};
-    border: ${(props) => (props.winner ? "0px" : "1px")} solid black;
-  }
-`;
-function hasWinner(x, y, board, dataSize) {
-  if (x === "" || y === "") return;
-  function columnHasWinner() {
-    let count = 0;
-    for (let i = 1; i < 5; i++) {
-      if (y + i >= dataSize) break;
-      if (board[y][x] === board[y + i][x] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    for (let i = 1; i < 5; i++) {
-      if (y - i < 0) break;
-      if (board[y][x] === board[y - i][x] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    if (count >= 4) return true;
-    return false;
-  }
-  function rowHasWinner() {
-    let count = 0;
-    for (let i = 1; i < 5; i++) {
-      if (x + i >= dataSize) break;
-      if (board[y][x] === board[y][x + i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    for (let i = 1; i < 5; i++) {
-      if (x - i < 0) break;
-      if (board[y][x] === board[y][x - i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    if (count >= 4) return true;
-    return false;
-  }
-  function backSlashHasWinner() {
-    let count = 0;
-    for (let i = 1; i < 5; i++) {
-      if (x + i >= dataSize || y + i >= dataSize) break;
-      if (board[y][x] === board[y + i][x + i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    for (let i = 1; i < 5; i++) {
-      if (x - i < 0 || y - i < 0) break;
-      if (board[y][x] === board[y - i][x - i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    if (count >= 4) return true;
-    return false;
-  }
-  function slashHasWinner() {
-    let count = 0;
-    for (let i = 1; i < 5; i++) {
-      if (x + i >= 20 || y - i < 0) break;
-      if (board[y][x] === board[y - i][x + i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    for (let i = 1; i < 5; i++) {
-      if (x - i < 0 || y + i >= dataSize) break;
-      if (board[y][x] === board[y + i][x - i] && board[y][x] !== null) {
-        count++;
-      } else {
-        break;
-      }
-    }
-    if (count >= 4) return true;
-    return false;
-  }
-  if (
-    rowHasWinner() ||
-    columnHasWinner() ||
-    backSlashHasWinner() ||
-    slashHasWinner()
-  )
-    return true;
-  return false;
-}
+import hasWinner from "./hasWinner";
+import { Container, Block, Title, StateInfo, ClickBlock } from "./components";
 
 let blackIsNext = true;
 function App() {
@@ -158,7 +20,6 @@ function App() {
       setNextInfo(winner === "black" ? "黑子贏了！" : "白子贏了！");
     }
   }, [board, x, y, dataSize, winner]);
-  useEffect(() => {}, [board]);
 
   function handleClick(indexX, indexY, value) {
     if (value !== null) return;
