+++ homeworks/week21/hw2/hw2/src/components/index.js
@@ -1,0 +1,2 @@
+import { Container, Block, Title, StateInfo, ClickBlock } from "./components";
+export {Container, Block, Title, StateInfo, ClickBlock}
