+++ homeworks/week21/hw2/hw2/src/hasWinner/index.js
@@ -1,0 +1,2 @@
+import hasWinner from "./hasWinner"
+export default hasWinner
