+++ homeworks/week21/hw4.md
@@ -1,7 +1,27 @@
 ## 為什麼我們需要 React？可以不用嗎？
+當網頁越來越龐大，元素越來越多的時候，要怎麼方便又有效率的處理這些任務，會靠 React 或其他框架處理這些問題。讓 code 更好維護，也降低瀏覽器需要重新渲染的東西。
+### React 的特色
+在學到 React 以前，我們都是直接操作 DOM 元素。在 React 中，會產生 Virtual DOM，來比對真正被改變的部份，再去改變真正的 DOM。可以降低重新渲染的成本，避免無謂的浪費資源。
+
+---
+
+React 裡面可以自定義組件（component），讓開發者可以把頁面上的元素氛圍更小塊的部份。這些組件可以重複利用，需要的時候不需要重寫一個。甚至也可以拿到別的專案使用。
+
+---
+
+React 是 Single-Way data flow，資料的流動是單一的。改變資料 ->改變畫面。開發者不用擔心畫面跟資料不一樣。React 就是改資料，畫面就會做相應的改變。
+
+---
+
+搜尋資料後發現，中小型專案不一定要用 React，可能用其他工具（像 vue）或是用原來的方式會比較快速，
 
 ## React 的思考模式跟以前的思考模式有什麼不一樣？
 
+React 都是操作資料，從而改變畫面。如果資料不變，畫面也不會改變。不會像之前，不僅要修改儲存的資料，也要修改畫面。在 React 裡面，都是透過操作 React 裡面儲存的資料。當 React 的資料改變， React 會用 virtual DOM 去比對改變的部份，並將那個部分重新渲染。降低瀏覽器的負擔。
 
 ## state 跟 props 的差別在哪裡？
+* props是React父元件與子元件間溝通的橋樑。
+* state 是元件的狀態。props 則是外部傳進元件的參數。
+* state 可以透過 setState 改變，而 props 是不可改變的。
+
